web1/requirements.txt
Django == 2.0.3
PyYAML == 5.1.2
requests == 2.18.4
eventlet == 0.22.1
mysqlclient == 1.4.4
django-celery == 3.2.2
flower == 0.9.2
dwebsocket == 0.4.2
paramiko == 2.4.1
HttpRunner == 2.2.5
PyMySQL == 0.9.3
python-ldap == 3.2.0
PyExecJS == 1.5.1
concurrent_log_handler == 0.9.16

